requirements.txt
Flask==2.1.3
numpy==1.22
tensorflow-cpu==2.12.0
opencv-python-headless==4.8.0.76
gunicorn==20.0.4
h5py==3.7.0
Keras==2.12.0





